Add RSS feed
Publish an RSS feed of the changes to the specs repository to make the
specs more discoverable.

Do not include the template in the toctree so it will not be included in
the RSS feed.

Use the doc build date to produce the copyright string instead of using
a hard-coded value.

Change-Id: Ib4f84efab8b3af3ddd47d38d63a91362612bd655